requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
python-dotenv==1.0.0
line-bot-sdk==3.8.0
pandas==2.2.0
psycopg2-binary==2.9.9
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6 